refactor: redefine the module structure
modules now use the event emitter for messaging, also include a method for sending messages to the background
