+++ entrypoints/content/types/bg-message.ts
@@ -4,5 +4,5 @@ export enum BgMessageId {
 
 export interface BgMessage {
     id: BgMessageId;
-    data: any;
+    data?: any;
 }
