+++ entrypoints/content/types/module.ts
@@ -1,13 +1,24 @@
+import {BgMessage} from "@/entrypoints/content/types/bg-message.ts";
+import EventEmitter from "eventemitter3";
+
 export enum ModuleId {
     PasswordChecker,
 }
 
+export abstract class Module extends EventEmitter {
+    abstract readonly id: ModuleId;
 
-export interface Module {
-    readonly id: ModuleId;
-    isEnabled: boolean;
+    abstract load(): void;
 
-    load(): void;
+    abstract unload(): void;
 
-    unload(): void;
+    protected sendToBackground(message: BgMessage): any {
+        browser.runtime.sendMessage(message, (response) => {
+            if (browser.runtime.lastError) {
+                console.error("Error sending message:", browser.runtime.lastError);
+            } else {
+                return response;
+            }
+        });
+    }
 }
